fix(ReactGridLayout): ensure window exists before checking navigator (#1057)
This was an issue for my usage as the package is imported in a node context, and this navigator check
is done at the top level of this package rather than after render the grid. So `window` is undefined
at the point it is initially included.

So we didn't even need to render the grid layout to have this issue.