fix(build): reduce npm package size by 30% by removing source in source maps
Given that the original source is present in the package, it can still be loaded
as-needed from the filesystem, so this was pure duplication. For a pure web build
without a bundler (who is still doing that?), if source maps are still needed,
this will possibly cause an issue. In that case we can look into committing a different
build file for umd.